+++ src/api/chats.controller.ts
@@ -1,11 +1,26 @@
-import { Controller, Delete, Get, Param, Post, Query } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Delete,
+  Get,
+  Param,
+  Post,
+  Put,
+  Query,
+} from '@nestjs/common';
 import { ApiOperation, ApiSecurity, ApiTags } from '@nestjs/swagger';
 
 import { SessionManager } from '../core/abc/manager.abc';
 import { WhatsappSession } from '../core/abc/session.abc';
 import { parseBool } from '../helpers';
 import { GetChatMessagesQuery } from '../structures/chats.dto';
-import { SessionApiParam, SessionParam } from './helpers';
+import { EditMessageRequest } from '../structures/chatting.dto';
+import {
+  ChatIdApiParam,
+  MessageIdApiParam,
+  SessionApiParam,
+  SessionParam,
+} from './helpers';
 
 @ApiSecurity('api_key')
 @Controller('api/:session/chats')
@@ -50,6 +65,33 @@ class ChatsController {
     @Param('chatId') chatId: string,
   ) {
     return session.clearMessages(chatId);
+  }
+
+  @Delete(':chatId/messages/:messageId')
+  @SessionApiParam
+  @ChatIdApiParam
+  @MessageIdApiParam
+  @ApiOperation({ summary: 'Deletes a message from the chat' })
+  deleteMessage(
+    @SessionParam session: WhatsappSession,
+    @Param('chatId') chatId: string,
+    @Param('messageId') messageId: string,
+  ) {
+    return session.deleteMessage(chatId, messageId);
+  }
+
+  @Put(':chatId/messages/:messageId')
+  @SessionApiParam
+  @ChatIdApiParam
+  @MessageIdApiParam
+  @ApiOperation({ summary: 'Edits a message in the chat' })
+  editMessage(
+    @SessionParam session: WhatsappSession,
+    @Param('chatId') chatId: string,
+    @Param('messageId') messageId: string,
+    @Body() body: EditMessageRequest,
+  ) {
+    return session.editMessage(chatId, messageId, body);
   }
 }